requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
datasets==2.3.2
dill==0.3.5.1
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
huggingface-hub==0.7.0
idna==3.3
multidict==6.0.2
multiprocess==0.70.13
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyarrow==8.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
responses==0.18.0
six==1.16.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
xxhash==3.0.0
yapf==0.32.0
yarl==1.7.2
